requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.1
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
passlib==1.7.4
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
